flask_api/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
langchain-ollama==0.1.0
langchain-core==0.1.52
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1
requests==2.31.0
google-generativeai==0.8.3
